♻️ refactor: authorized コールバックを edge.ts に分離 #59
- config.ts から authorized を削除（Edge専用のため）
- edge.ts に authorized コールバックを追加
- proxy.ts を edge.ts からインポートするように変更
- docs/authentication.md を更新
